[core] Various fixes.

+++ core/idt.c
@@ -13,7 +13,7 @@ struct idt_entry {
 
 struct idt_ptr {
 	unsigned short limit;
-	unsigned int base;
+	unsigned long base; /* int? */
 } __attribute__((packed));
 
 struct idt_entry idt[256];
@@ -46,7 +46,7 @@ idt_set_gate(
 void
 idt_install() {
 	idtp.limit = (sizeof(struct idt_entry) * 256) - 1;
-	idtp.base = &idt;
+	idtp.base = (uintptr_t)&idt;
 	memset(&idt, 0, sizeof(struct idt_entry) * 256);
 
 	idt_load();
